Machine having different immediate widths, and only smaller of them connected to some port, and no regcopyadder needed, resulted in failing schedule.
Now checks also for connectivity when checking if short immediate can be scheduled.

This also makes the scheduler use more long immediates instead of short immediates 
and regcopyadder on some other machines, affecting performance.

Performance is much faster with sparsely-connected machines with lots of busses, 
worse with machines with reduced connectivity and few number of buses, 
all of them not being available when limm in use