Implemented support for "guard pipeline"  in compiled sim;
Compiled simulator can now simulate correctly machines which have longer guard latency of 1,
(if it's global 1 + RF N; global guard latency >2 still not supported).

So the simulator no longer reverts to interpreting mode when simulating these machines.

This should speed up simulation of some high-clock-speed architechtures by a huge margin.


The guard pipeline is disabled if it's not needed so it does not make simulation of machines
with guard latency of only 1 any slower.