Implemented support for SPU's branch hint instructions.
To make them work, the DDG builder now uses the (optional) ProgramOperations
created during LLVM conversion. These are created only if functionAtATime
is enabled. They are not created in case of the regular TCE mode because
the POs will get invalid after the Program is copied as they will end up
having references to the Moves of the previous POM. 

ProgramOperations are now handled via boost::shared_ptr references to avoid
ownership problems. They are created in LLVMTCEBuilder, set to TerminalFUPorts and
then reused in DDG building. DDG takes ownership of the MoveNodes inside the POs
as it's not so easy to apply shared_ptrs for MoveNodes due to them being part
of the Graph framework.