refactoring/cleanup:
UniversalMachine is now a singleton and used solely for marking
unassigned Move resources. Call UniversalMachine::instance() to
get access to an instance, do not pass it around to methods.

Some APIs require still it in the method argument list as their
behavior is changed whether it's given or not. Get rid of these
gradually.