Port FileSystem class to Boost Filesystem version 3
Most of this work was done very directly accroding to the table at
http://www.boost.org/doc/libs/1_52_0/libs/filesystem/doc/deprecated.html

Since the boost::filesystem::path class does not support the custom
path validity checker function, it was entirely left out. An alternative
would have been to keep the path checker function but hook it to some
other mechanism.