Restore compatibility with Boost filesystem v2
Since TCE currently targets platforms that do not easily have access to
boost filesystem v3, this change tries to use older v2 APIs instead of
ones that are missing in v3.

The selection on which API to compile against depends on the
BOOST_FILESYSTEM_VERSION macro, so it should always be set in
FileSystem.hh as long as these ifdef blocks are kept.