Made the 'lazy' and 'local' settings of dlopen() in PluginTools explicit. Each client now sets those explicitly.
base.opb links against libtce.so to make it work through an icd loader and
pocl.

icd loaders have RTLD_LOCAL set when they call dlopen for the ocl
library. In our case it's libpocl.so which loads libtce.so which later on 
loads xxxx.opb that fails in case xxxx.opb needs a symbol from libtce.so that 
wasn't imported by libpocl.so.

Still needs fixing for the custom operations: they should have the -ltce
dependency.