Ensure libtce.so in the build dir is tested, not an installed one.
The current fix to https://github.com/cpc/tce/issues/91 breaks
build dir execution when one has an installed libtce.so. This
is because it makes LD_LIBRARY_PATH the preference (which could
point to a custom library location with a libtce.so) instead
of the build dir one, like the libtool wrapper is expected
to do.