Add a separate developer mode enabled via TCE_DEVEL_MODE=1
Developers should now set TCE_DEVEL_MODE env to 1 to cause the
previous behavior of searching for build paths first.
compiletest.sh and systemtest.py add it automatically, so
it's needed to be added manually only in case running the build dir
binaries outside the tester scripts.

The previous semi-automated way to figure it out whether using an
installed or build dir TCE was flaky and simply did not work
when using libtce.so instead of the binaries. This problem was
exposed by pocl's ttasim driver.