functions_of_interest to the DSE
An application descriptor file that can be used to set the functions
of interest in the simulated program.  If given, only the time spent in
the chosen functions is included in the execution time that guides
the design space exploration.

This is useful for cases where the main() contains printouts and verification
code which then launches the actual function to optimize for.